Properly kill the X used for authentication v3
This is to fix MeeGo Bug #67: EFS can not login the system with
correct password, which is caused by a race condition.

When EFS is enabled, uxlaunch starts an Xorg instance for
authentication. After that, it was killed, but uxlaunch doesn't
waitfor it to be ended before starting the second Xorg instance, so
there are chances that the second Xorg can't start due to the first X
hasn't finished the cleaning up. This patch fixed this issue.

v3: no need to add a new global function, we can kill X in place

v2: stop dbus session bus

Signed-off-by: Yan Li <yan.i.li@intel.com>